Refactor: Complete refactoring

+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Identity.API/Controllers/UserController.cs
@@ -29,27 +29,30 @@ public class UserController(IUserService _userService, ILogger<UserController> _
     [Authorize(Roles = UserConstant.ADMIN_ROLE)]
     public async Task<ApiResponse<bool>> DeleteById(long id)
     {
+        if (EnvUtil.IsProduction())
+            throw new ApiException(ErrorCode.PARAMS_ERROR, "User delete is disabled in production environment");
         if (id == 0) throw new ApiException(ErrorCode.PARAMS_ERROR, "Invalid user id");
 
         await _userService.DeleteAsync(id);
         return ResultUtil.Success(true);
     }
 
-    // Disable edit user endpoint, edit in database directly
-    // [HttpPost]
-    // [Authorize(Roles = UserConstant.ADMIN_ROLE)]
-    // public async Task<ApiResponse<bool>> UpdateByRequest([FromBody] UserUpdateRequest request)
-    // {
-    //     if (request == null)
-    //         return ResultUtil.Error<bool>(ErrorCode.PARAMS_ERROR, "Invalid user");
-
-    //     User? user = await _userService.GetByIdAsTrackingAsync(request.Id);
-    //     if (user == null)
-    //         return ResultUtil.Error<bool>(ErrorCode.PARAMS_ERROR, "User not found");
-
-    //     _mapper.Map(request, user);
-    //     await _userService.UpdateAsync(user);
-    //     return ResultUtil.Success(true);
-    // }
+    [HttpPost]
+    [Authorize(Roles = UserConstant.ADMIN_ROLE)]
+    public async Task<ApiResponse<bool>> UpdateByRequest([FromBody] UserUpdateRequest request)
+    {
+        var user = await _userService.GetByIdAsync(request.Id);
+        if (user == null) throw new ApiException(ErrorCode.NOT_FOUND_ERROR, "User not found");
+        if (!string.IsNullOrWhiteSpace(request.UserEmail))
+            user.UserEmail = request.UserEmail;
+        if (!string.IsNullOrWhiteSpace(request.UserName))
+            user.UserName = request.UserName;
+        if (!string.IsNullOrWhiteSpace(request.UserAvatar))
+            user.UserAvatar = request.UserAvatar;
+        if (!string.IsNullOrWhiteSpace(request.UserProfile))
+            user.UserProfile = request.UserProfile;
+        await _userService.UpdateAsync(user);
+        return ResultUtil.Success(true);
+    }
 
 }
